chatbot/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.40,<2.1
aiosqlite==0.19.0
PyMySQL==1.1.1
cryptography>=42.0.0
httpx==0.26.0
python-multipart==0.0.22
pydantic[email]
python-dotenv==1.0.1
